requirements.txt
sqlalchemy-iris
langchain-iris
testcontainers-iris
llama-iris
sentence-transformers
langchain
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel
setuptools
gradio_client
moviepy
ffmpeg-python
replicate
urllib3>=2.0.0a1
streamlit
geopy